+++ tests/test_fold_inpainting.py
@@ -2,9 +2,10 @@ import json
 import sys
 from pathlib import Path
 
-from rush.boltz import ProteinSequence, boltz
-from rush.client import RunOpts, RunSpec, set_opts
+from rush.boltz import BoltzSavedResult, ProteinSequence, boltz, save_outputs
+from rush.client import RunError, RunOpts, RunSpec, set_opts
 from rush.mmseqs2 import mmseqs2
+from rush.mmseqs2 import save_outputs as save_mmseqs2_outputs
 
 
 def test_fold_inpainting():
@@ -19,7 +20,11 @@ def test_fold_inpainting():
         ),
         collect=True,
     )
-    print(res, file=sys.stderr)
+    assert not isinstance(res, (str, RunError))
+    print(json.dumps(res, indent=2), file=sys.stderr)
+    saved_msas = save_mmseqs2_outputs(res)
+    assert not isinstance(saved_msas, (str, RunError))
+    assert saved_msas[0].suffix == ".a3m"
     res = boltz(
         [
             ProteinSequence(["A"], protein_seq, res[0]),
@@ -34,7 +39,11 @@ def test_fold_inpainting():
         run_spec=RunSpec(target="Bullet", gpus=1),
         collect=True,
     )
+    assert not isinstance(res, (str, RunError))
     print(json.dumps(res, indent=2), file=sys.stderr)
+    output = save_outputs(res)
+    assert not isinstance(output, (str, RunError))
+    assert isinstance(output[0], BoltzSavedResult)
 
 
 if __name__ == "__main__":